Make MarkdownEditor more abstract
