+++ markdown/src/main/java/it/niedermann/android/markdown/MarkdownEditor.java
@@ -1,82 +1,13 @@
 package it.niedermann.android.markdown;
 
-import android.annotation.SuppressLint;
-import android.content.Context;
-import android.util.AttributeSet;
-import android.webkit.JavascriptInterface;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
-
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.core.util.Consumer;
 
-public class MarkdownEditor extends WebView {
+public interface MarkdownEditor {
 
-    private boolean pageFinished = false;
-    @NonNull
-    private Consumer<String> listener = s -> {
-    };
-    private String textToSetOnPageFinished;
-    private boolean enabledStateOnPageFinished = true;
+    void setText(String textToSetOnPageFinished);
 
-    @SuppressLint("SetJavaScriptEnabled")
-    public MarkdownEditor(@NonNull Context context) {
-        super(context);
-        init();
-    }
+    void setTextChangedListener(@NonNull Consumer<String> listener);
 
-    @SuppressLint({"SetJavaScriptEnabled", "JavascriptInterface", "AddJavascriptInterface" })
-    public MarkdownEditor(@NonNull Context context, @Nullable AttributeSet attrs) {
-        super(context, attrs);
-        init();
-    }
-
-    @SuppressLint("SetJavaScriptEnabled")
-    public MarkdownEditor(@NonNull Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
-        super(context, attrs, defStyleAttr);
-        init();
-    }
-
-    @SuppressLint("SetJavaScriptEnabled")
-    private void init() {
-        getSettings().setJavaScriptEnabled(true);
-        addJavascriptInterface(this, "Android");
-        loadUrl("file:///android_asset/web/index.html");
-        setWebViewClient(new WebViewClient() {
-            @Override
-            public void onPageFinished(WebView view, String url) {
-                super.onPageFinished(view, url);
-                pageFinished = true;
-                setText(textToSetOnPageFinished);
-                setEnabled(enabledStateOnPageFinished);
-            }
-        });
-    }
-
-    public void setText(String textToSetOnPageFinished) {
-        if (pageFinished) {
-            String newText = this.textToSetOnPageFinished == null ? "" : this.textToSetOnPageFinished.replace("`", "\\`");
-            evaluateJavascript("setText(`" + newText + "`);", value -> {});
-        } else {
-            this.textToSetOnPageFinished = textToSetOnPageFinished;
-        }
-    }
-
-    public void setTextChangedListener(@NonNull Consumer<String> listener) {
-        this.listener = listener;
-    }
-
-    public void setEnabled(boolean enabled) {
-        if (pageFinished) {
-            evaluateJavascript("setEnabled(" + this.enabledStateOnPageFinished + ");", value -> {});
-        } else {
-            this.enabledStateOnPageFinished = enabled;
-        }
-    }
-
-    @JavascriptInterface
-    public void onTextChanged(String newText) {
-        this.listener.accept(newText);
-    }
+    void setEnabled(boolean enabled);
 }
